bean validator

+++ src/main/resources/db/migration/V0001__create_table_usuario.sql
@@ -1,5 +1,5 @@
 create table tb_usuario(
-	id bigint primary key,	
+	id serial primary key,	
 	nome varchar(200) not null,
 	idade smallint
 );
